+++ lib/wechat/shake_around.rb
@@ -8,8 +8,13 @@ require 'wechat/shake_around/material'
 require 'wechat/shake_around/icon'
 require 'wechat/shake_around/license'
 
+require 'wechat/shake_around/page'
+require 'wechat/shake_around/beacon_page_relation'
+
+require 'wechat/shake_around/page_report'
+require 'wechat/shake_around/page_daily_report'
+
 module Wechat
   module ShakeAround
-    # Your code goes here...
   end
 end